Feature ag 3570

See merge request choerodon-agile/agile-service!583

+++ src/main/java/io/choerodon/agile/infra/dataobject/GroupDataChartListDO.java
@@ -4,6 +4,8 @@ import io.choerodon.agile.api.dto.IssueTypeDTO;
 import io.choerodon.agile.api.dto.PriorityDTO;
 import io.choerodon.agile.api.dto.StatusMapDTO;
 
+import java.math.BigDecimal;
+
 /**
  * Created by HuangFuqiang@choerodon.io on 2018/7/28.
  * Email: fuqianghuang01@gmail.com
@@ -22,9 +24,9 @@ public class GroupDataChartListDO {
 
     private String statusName;
 
-    private Integer storyPoints;
+    private BigDecimal storyPoints;
 
-    private Integer remainTime;
+    private BigDecimal remainTime;
 
     private int completed;
 
@@ -90,19 +92,19 @@ public class GroupDataChartListDO {
         this.statusName = statusName;
     }
 
-    public void setStoryPoints(Integer storyPoints) {
+    public void setStoryPoints(BigDecimal storyPoints) {
         this.storyPoints = storyPoints;
     }
 
-    public Integer getStoryPoints() {
+    public BigDecimal getStoryPoints() {
         return storyPoints;
     }
 
-    public void setRemainTime(Integer remainTime) {
+    public void setRemainTime(BigDecimal remainTime) {
         this.remainTime = remainTime;
     }
 
-    public Integer getRemainTime() {
+    public BigDecimal getRemainTime() {
         return remainTime;
     }
